Extract methods for running MachineShop
Moved simulate, runSimulatation, and main
from MachineShopSimulator into a new class
called RunMachineShop. Also moved numjobs
there. Tests are still passing.
